Added baseclass for user and changed other classes to use use UserInterface instead of user entity
