+++ src/Sulu/Bundle/SecurityBundle/Entity/RoleInterface.php
@@ -86,30 +86,30 @@ interface RoleInterface
     /**
      * Set creator
      *
-     * @param \Sulu\Bundle\SecurityBundle\Entity\User $creator
+     * @param \Sulu\Component\Security\UserInterface $creator
      * @return Role
      */
-    public function setCreator(\Sulu\Bundle\SecurityBundle\Entity\User $creator = null);
+    public function setCreator(\Sulu\Component\Security\UserInterface $creator = null);
 
     /**
      * Get creator
      *
-     * @return \Sulu\Bundle\SecurityBundle\Entity\User
+     * @return \Sulu\Component\Security\UserInterface
      */
     public function getCreator();
 
     /**
      * Set changer
      *
-     * @param \Sulu\Bundle\SecurityBundle\Entity\User $changer
+     * @param \Sulu\Component\Security\UserInterface $changer
      * @return Role
      */
-    public function setChanger(\Sulu\Bundle\SecurityBundle\Entity\User $changer = null);
+    public function setChanger(\Sulu\Component\Security\UserInterface $changer = null);
 
     /**
      * Get changer
      *
-     * @return \Sulu\Bundle\SecurityBundle\Entity\User
+     * @return \Sulu\Component\Security\UserInterface
      */
     public function getChanger();
