+++ src/Sulu/Bundle/SecurityBundle/Entity/User.php
@@ -1,53 +1,22 @@
 <?php
-/*
-* This file is part of the Sulu CMS.
-*
-* (c) MASSIVE ART WebServices GmbH
-*
-* This source file is subject to the MIT license that is bundled
-* with this source code in the file LICENSE.
-*/
 
 namespace Sulu\Bundle\SecurityBundle\Entity;
 
-use Serializable;
-use JMS\Serializer\Annotation\ExclusionPolicy;
 use JMS\Serializer\Annotation\Expose;
-use JMS\Serializer\Annotation\VirtualProperty;
-use JMS\Serializer\Annotation\SerializedName;
-use Sulu\Bundle\CoreBundle\Entity\ApiEntity;
-use Sulu\Component\Security\UserInterface;
+use JMS\Serializer\Annotation\ExclusionPolicy;
 
 /**
  * User
  *
  * @ExclusionPolicy("all")
  */
-class User extends ApiEntity implements UserInterface, Serializable
+class User extends BaseUser
 {
-    /**
-     * @var string
-     * @Expose
-     */
-    private $username;
-
-    /**
-     * @var string
-     * @Expose
-     */
-    private $password;
-
-    /**
-     * @var string
-     * @Expose
-     */
-    private $locale;
-
     /**
      * @var integer
      * @Expose
      */
-    private $id;
+    private $userLevel = 0;
 
     /**
      * @var \Sulu\Bundle\ContactBundle\Entity\Contact
@@ -62,17 +31,6 @@ class User extends ApiEntity implements UserInterface, Serializable
     private $userRoles;
 
     /**
-     * @var string
-     */
-    private $salt;
-
-    /**
-     * @var string
-     * @Expose
-     */
-    private $privateKey;
-
-    /**
      * @var \Doctrine\Common\Collections\Collection
      * @Expose
      */
@@ -84,153 +42,26 @@ class User extends ApiEntity implements UserInterface, Serializable
     private $userSettings;
 
     /**
-     * @var string
-     */
-    private $apiKey;
-
-    /**
-     * @var boolean
-     */
-    private $locked = false;
-
-    /**
-     * @var boolean
-     * @Expose
-     */
-    private $enabled = true;
-
-    /**
-     * @var \DateTime
-     */
-    private $lastLogin;
-
-    /**
-     * @var string
-     */
-    private $confirmationKey;
-
-    /**
-     * Constructor
-     */
-    public function __construct()
-    {
-        $this->userRoles = new \Doctrine\Common\Collections\ArrayCollection();
-        $this->userGroups = new \Doctrine\Common\Collections\ArrayCollection();
-        $this->userSettings = new \Doctrine\Common\Collections\ArrayCollection();
-
-        $this->apiKey = md5(uniqid());
-    }
-
-    /**
-     * Set username
+     * Set userLevel
      *
-     * @param string $username
+     * @param integer $userLevel
      * @return User
      */
-    public function setUsername($username)
+    public function setUserLevel($userLevel)
     {
-        $this->username = $username;
+        $this->userLevel = $userLevel;
 
         return $this;
     }
 
     /**
-     * Get username
-     *
-     * @return string
-     */
-    public function getUsername()
-    {
-        return $this->username;
-    }
-
-    /**
-     * @VirtualProperty
-     * @SerializedName("fullName")
-     * @return string
-     */
-    public function getFullName()
-    {
-        return $this->getContact()->getFullName();
-    }
-
-    /**
-     * Set password
-     *
-     * @param string $password
-     * @return User
-     */
-    public function setPassword($password)
-    {
-        $this->password = $password;
-
-        return $this;
-    }
-
-    /**
-     * Get password
-     *
-     * @return string
-     */
-    public function getPassword()
-    {
-        return $this->password;
-    }
-
-    /**
-     * Set locale
-     *
-     * @param string $locale
-     * @return User
-     */
-    public function setLocale($locale)
-    {
-        $this->locale = $locale;
-
-        return $this;
-    }
-
-    /**
-     * Get locale
-     *
-     * @return string
-     */
-    public function getLocale()
-    {
-        return $this->locale;
-    }
-
-    /**
-     * Get id
+     * Get userLevel
      *
      * @return integer
      */
-    public function getId()
+    public function getUserLevel()
     {
-        return $this->id;
-    }
-
-    /**
-     * Set contact
-     *
-     * @param \Sulu\Bundle\ContactBundle\Entity\Contact $contact
-     * @return User
-     */
-    public function setContact(\Sulu\Bundle\ContactBundle\Entity\Contact $contact = null)
-    {
-        $this->contact = $contact;
-
-        return $this;
-    }
-
-    /**
-     * Get contact
-     *
-     * @return \Sulu\Bundle\ContactBundle\Entity\Contact
-     */
-    public function getContact()
-    {
-        return $this->contact;
+        return $this->userLevel;
     }
 
     /**
@@ -264,96 +95,6 @@ class User extends ApiEntity implements UserInterface, Serializable
     public function getUserRoles()
     {
         return $this->userRoles;
-    }
-
-    /**
-     * Set salt
-     *
-     * @param string $salt
-     * @return User
-     */
-    public function setSalt($salt)
-    {
-        $this->salt = $salt;
-
-        return $this;
-    }
-
-    /**
-     * Get salt
-     *
-     * @return string
-     */
-    public function getSalt()
-    {
-        return $this->salt;
-    }
-
-    /**
-     * Set privateKey
-     *
-     * @param string $privateKey
-     * @return User
-     */
-    public function setPrivateKey($privateKey)
-    {
-        $this->privateKey = $privateKey;
-
-        return $this;
-    }
-
-    /**
-     * Get privateKey
-     *
-     * @return string
-     */
-    public function getPrivateKey()
-    {
-        return $this->privateKey;
-    }
-
-    /**
-     * Returns just the default symfony user role, so that the user get recognized as authenticated by symfony
-     *
-     * @return array The user roles
-     */
-    public function getRoles()
-    {
-        return array('ROLE_USER');
-    }
-
-    /**
-     * Removes the password of the user
-     *
-     * @return void
-     */
-    public function eraseCredentials()
-    {
-    }
-
-    /**
-     * Serializes the user just with the id, as it is enough
-     * @link http://php.net/manual/en/serializable.serialize.php
-     * @return string The string representation of the object or null
-     */
-    public function serialize()
-    {
-        return serialize(
-            array(
-                $this->id
-            )
-        );
-    }
-
-    /**
-     * Constructs the object
-     * @link http://php.net/manual/en/serializable.unserialize.php
-     * @param string $serialized The string representation of the object.
-     * @return void
-     */
-    public function unserialize($serialized)
-    {
-        list ($this->id) = unserialize($serialized);
     }
 
     /**
@@ -422,117 +163,35 @@ class User extends ApiEntity implements UserInterface, Serializable
     }
 
     /**
-     * Set apiKey
+     * Set contact
      *
-     * @param string $apiKey
+     * @param \Sulu\Bundle\ContactBundle\Entity\Contact $contact
      * @return User
      */
-    public function setApiKey($apiKey)
+    public function setContact(\Sulu\Bundle\ContactBundle\Entity\Contact $contact = null)
     {
-        $this->apiKey = $apiKey;
+        $this->contact = $contact;
 
         return $this;
     }
 
     /**
-     * Get apiKey
+     * Get contact
      *
+     * @return \Sulu\Bundle\ContactBundle\Entity\Contact
+     */
+    public function getContact()
+    {
+        return $this->contact;
+    }
+
+    /**
+     * @VirtualProperty
+     * @SerializedName("fullName")
      * @return string
      */
-    public function getApiKey()
+    public function getFullName()
     {
-        return $this->apiKey;
-    }
-
-    /**
-     * Set locked
-     *
-     * @param boolean $locked
-     * @return User
-     */
-    public function setLocked($locked)
-    {
-        $this->locked = $locked;
-
-        return $this;
-    }
-
-    /**
-     * Get locked
-     *
-     * @return boolean
-     */
-    public function getLocked()
-    {
-        return $this->locked;
-    }
-
-    /**
-     * Set enabled
-     *
-     * @param boolean $enabled
-     * @return User
-     */
-    public function setEnabled($enabled)
-    {
-        $this->enabled = $enabled;
-
-        return $this;
-    }
-
-    /**
-     * Get enabled
-     *
-     * @return boolean
-     */
-    public function getEnabled()
-    {
-        return $this->enabled;
-    }
-
-    /**
-     * Set lastLogin
-     *
-     * @param \DateTime $lastLogin
-     * @return User
-     */
-    public function setLastLogin($lastLogin)
-    {
-        $this->lastLogin = $lastLogin;
-
-        return $this;
-    }
-
-    /**
-     * Get lastLogin
-     *
-     * @return \DateTime
-     */
-    public function getLastLogin()
-    {
-        return $this->lastLogin;
-    }
-
-    /**
-     * Set confirmationKey
-     *
-     * @param string $confirmationKey
-     * @return User
-     */
-    public function setConfirmationKey($confirmationKey)
-    {
-        $this->confirmationKey = $confirmationKey;
-
-        return $this;
-    }
-
-    /**
-     * Get confirmationKey
-     *
-     * @return string
-     */
-    public function getConfirmationKey()
-    {
-        return $this->confirmationKey;
+        return $this->getContact()->getFullName();
     }
 }
